+++ packages/panels/src/components/nav-bar/NavBarSearchField.tsx
@@ -1,9 +1,7 @@
 import * as React from "react";
 import { TextInput, TextInputProps } from "@stenajs-webui/forms";
 import { Clickable } from "@stenajs-webui/core";
-import { Icon } from "@stenajs-webui/elements";
-import { faSearch } from "@fortawesome/free-solid-svg-icons/faSearch";
-import { cssColor } from "@stenajs-webui/theme";
+import { Icon, stenaSearch } from "@stenajs-webui/elements";
 import cx from "classnames";
 import styles from "./NavBarSearchField.module.css";
 import { useNavBarVariant } from "./NavBarVariantContext";
@@ -18,15 +16,11 @@ export const NavBarSearchField: React.FC<NavBarSearchFieldProps> = ({
   const variant = useNavBarVariant();
   return (
     <TextInput
-      wrapperClassName={cx(
-        styles.navBarSearchField,
-        styles[variant],
-        className
-      )}
+      className={cx(styles.navBarSearchField, styles[variant], className)}
       placeholder={placeholder}
       contentRight={
         <Clickable>
-          <Icon icon={faSearch} color={cssColor("--lhds-color-ui-50")} />
+          <Icon icon={stenaSearch} />
         </Clickable>
       }
       {...textInputProps}